refactor: setup implementation decoupling
