+++ app/repositories/setup_impl/README.md
@@ -1,0 +1,2 @@
+### Setup
+Since the setup is not highly specific to the underlying platform, we don't provide an abstraction to it like in the other parts of the framework. Instead, the platform should provide a router that is simply attached to the main router in `main.py` by the API.
